feat: profile page, part 2/n
