+++ src/app/common/components/app-bar/app-bar.component.scss
@@ -3,7 +3,7 @@
 @use "../../../../assets/mixins" as *;
 
 div.appBar {
-    padding-block: 0.5rem;
+    padding-block: 0.8rem;
     width: 100%;
     display: flex;
     justify-content: space-between;
@@ -49,7 +49,6 @@ div.appBar {
         @media only screen and (max-width: 686px) and (min-width: 420px) {
             #app_title_full {
                 display: initial;
-                font-size: 1rem !important;
             }
             #user_name {
                 display: none;
