+++ src/app/common/components/bottom-app-bar/bottom-app-bar.component.html
@@ -153,7 +153,8 @@
             With <fa-icon [icon]="heartIcon"></fa-icon> by
             <a href="https://cambo.in">Maninderpal Singh</a>
             <br /><br />
-            &copy; Cambo {{ getYear() }}
+            <fa-icon [icon]="copyIcon" size="xs"></fa-icon> Cambo
+            {{ getYear() }}
         </div>
     </div>
