+++ src/app/editor/components/dashboard-card/dashboard-card.component.scss
@@ -4,7 +4,7 @@
     @include basicBack;
     width: 186px;
     // opacity: 0.9;
-    border-radius: 2rem;
+    border-radius: 1rem;
     margin: 0.2rem;
 
     div {
