+++ src/app/editor/pages/editor-menu/editor-menu.component.scss
@@ -45,62 +45,6 @@
     border-radius: 24px;
 }
 
-.account-field {
-    .field-row {
-        display: flex;
-        flex-wrap: wrap;
-        align-items: center;
-        justify-content: space-between;
-        margin-block: 0.3rem;
-        font-size: 1rem;
-
-        .label {
-            margin-top: 0.5rem;
-            margin-right: 1rem;
-        }
-
-        input {
-            background-color: whitesmoke;
-            border-radius: 1rem;
-            padding: 8px;
-            padding-inline: 12px;
-            border-width: 0px;
-            font-size: 1rem;
-
-            @include sweetBorder;
-        }
-    }
-    .field-note {
-        font-size: 0.6rem;
-        margin-bottom: 1rem;
-    }
-
-    .name-preview {
-        border-radius: 0.5rem;
-        font-size: 0.6rem;
-        padding: 0.2rem;
-        padding-inline: 0.4rem;
-        margin-left: 0.3rem;
-        @include sweetBorder;
-    }
-}
-#profile_save_btn {
-    background-color: whitesmoke;
-    border-radius: 1rem;
-    cursor: pointer;
-    padding: 10px;
-    padding-inline: 12px;
-    margin-left: auto;
-    margin-top: 2rem;
-    width: 7rem;
-    text-align: center;
-    @include sweetBorder;
-}
-
-#profile_save_btn:hover {
-    background-color: white;
-}
-
 @media only screen and (max-width: 900px) {
     create-dialog {
         margin-inline: auto;
